+++ packages/core/src/generators/vue/index.ts
@@ -1,1 +1,2 @@
 export * from './vue';
+export * from './types';
